% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rda.R
\name{dmsg}
\alias{dmsg}
\title{Possibly print a debugging message}
\usage{
dmsg(debug, ...)
}
\arguments{
\item{debug}{integer (or logical, which will be converted to
an integer) that controls printing. If \code{debug}
this exceeds 0, then the \code{...} values are printed using \code{\link[=cat]{cat()}}.}

\item{...}{passed to \code{\link[=cat]{cat()}}, which prints the message.}
}
\description{
Possibly print a debugging message
}
\author{
Dan Kelley
}
