% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{removeRDA}
\alias{removeRDA}
\title{Remove a Value From an RDA file}
\usage{
removeRDA(name = NULL, rda = whichRDA())
}
\arguments{
\item{name}{character value indicating the name of the quantity
to be removed.}

\item{rda}{character value indicating the name of the RDA file in
which to save the notes. The default is the value returned by
\code{\link[=whichRDA]{whichRDA()}}, as set by a previous call to \code{\link[=useRDA]{useRDA()}}.}
}
\description{
This removes a value from an RDA notebook file.
}
\examples{
library(rdan)
rdaFile <- tempfile(pattern = "rdan", fileext = ".rda")
useRDA(rdaFile)
storeRDA("2pi", 2 * pi, "circle perimeter divided by radius")
retrieveRDA("2pi")
storeRDA("4pi", 4 * pi, "twice 2pi")
listRDA()
removeRDA("2pi")
listRDA()
removeRDA("4pi")
listRDA()
# Clean up
file.remove(rdaFile)

}
\author{
Dan Kelley
}
