% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{retrieveRDA}
\alias{retrieveRDA}
\title{Retrieve a Value from the RDA Notebook}
\usage{
retrieveRDA(name = NULL, rda = whichRDA())
}
\arguments{
\item{name}{character value naming the item, as set by \code{\link[=storeRDA]{storeRDA()}}.}

\item{rda}{character value indicating the name of the RDA file in
which to save the notes. The default is the value returned by
\code{\link[=whichRDA]{whichRDA()}}, as set by a previous call to \code{\link[=useRDA]{useRDA()}}.}
}
\value{
a list containing \code{name}, \code{value}, \code{description} and
\code{context}, as set by the previous call to \code{\link[=storeRDA]{storeRDA()}}.
}
\description{
This returns a list with items named \code{value},
\code{description}, and \code{context}, as set by \code{\link[=storeRDA]{storeRDA()}}.
}
\author{
Dan Kelley
}
