% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{storeRDA}
\alias{storeRDA}
\title{Store a Value Into an RDA file}
\usage{
storeRDA(
  name = NULL,
  value = NULL,
  comment = "",
  context = "",
  rda = whichRDA()
)
}
\arguments{
\item{name}{character value indicating the name of the quantity
to be stored. Note that this does not need to be a valid R
identifier (see Example).}

\item{value}{an R object storing the value to be stored.}

\item{comment}{character value holding a comment about the quantity
being stored.}

\item{context}{character value indicating something about the context
of the stored item. If this is an empty string (which is the default)
then it is the current working directory as discovered with \code{\link[=getwd]{getwd()}}.
If it is a string starting with the \code{/} character, then
the current working directory is prepended to the string.
Otherwise, `context' is used directly.}

\item{rda}{character value indicating the name of the RDA file in
which to save the notes. The default is the value returned by
\code{\link[=whichRDA]{whichRDA()}}, as set by a previous call to \code{\link[=useRDA]{useRDA()}}.}
}
\description{
This stores a value, its description, and its context within
an RDA notebook file.
}
\author{
Dan Kelley
}
