% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.R
\name{useRDA}
\alias{useRDA}
\title{Designate an RDA File for Notekeeping}
\usage{
useRDA(rda = "notes.rda")
}
\arguments{
\item{rda}{character value indicating the name of the RDA file in
which to save the notes. The default is the value returned by
\code{\link[=whichRDA]{whichRDA()}}, as set by a previous call to \code{\link[=useRDA]{useRDA()}}.}
}
\description{
This creates the RDA file if it does not exist yet. It also
has a side-effect by storing the file's path in a hidden
way that can be discovered by calling \code{\link[=whichRDA]{whichRDA()}}, which
is used by \code{\link[=retrieveRDA]{retrieveRDA()}} and \code{\link[=storeRDA]{storeRDA()}}.
}
\author{
Dan Kelley
}
