% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rda.R
\name{whichRDA}
\alias{whichRDA}
\title{Find name of RDA file}
\usage{
whichRDA()
}
\description{
Until \code{\link[=useRDA]{useRDA()}} is called, this will be \code{RDAN.rda}, but after
that, it is the value specified in the call to \code{\link[=useRDA]{useRDA()}}.
}
