% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdan.R
\name{dmsg}
\alias{dmsg}
\title{Possibly print a debugging message}
\usage{
dmsg(...)
}
\arguments{
\item{...}{passed to \code{\link[=cat]{cat()}}, which prints the message.}
}
\description{
Print a debugging message, if \code{\link[=debugLevel]{debugLevel()}} returns TRUE,
i.e. if \code{\link[=debugOn]{debugOn()}} has been called, without a subsequent
call to \code{\link[=debugOff]{debugOff()}}.
}
\author{
Dan Kelley
}
